+++ src/Storageless/Http/SessionMiddleware.php
@@ -18,7 +18,7 @@
 
 declare(strict_types=1);
 
-namespace PSR7Session\Http;
+namespace PSR7Sessions\Storageless\Http;
 
 use Dflydev\FigCookies\FigResponseCookies;
 use Dflydev\FigCookies\SetCookie;
@@ -29,11 +29,11 @@ use Lcobucci\JWT\Token;
 use Lcobucci\JWT\ValidationData;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
-use PSR7Session\Time\CurrentTimeProviderInterface;
-use PSR7Session\Time\SystemCurrentTime;
-use PSR7Session\Session\DefaultSessionData;
-use PSR7Session\Session\LazySession;
-use PSR7Session\Session\SessionInterface;
+use PSR7Sessions\Storageless\Time\CurrentTimeProviderInterface;
+use PSR7Sessions\Storageless\Time\SystemCurrentTime;
+use PSR7Sessions\Storageless\Session\DefaultSessionData;
+use PSR7Sessions\Storageless\Session\LazySession;
+use PSR7Sessions\Storageless\Session\SessionInterface;
 use Zend\Stratigility\MiddlewareInterface;
 
 final class SessionMiddleware implements MiddlewareInterface
@@ -248,12 +248,13 @@ final class SessionMiddleware implements MiddlewareInterface
     private function appendToken(SessionInterface $sessionContainer, Response $response, Token $token = null) : Response
     {
         $sessionContainerChanged = $sessionContainer->hasChanged();
+        $sessionContainerEmpty   = $sessionContainer->isEmpty();
 
-        if ($sessionContainerChanged && $sessionContainer->isEmpty()) {
+        if ($sessionContainerChanged && $sessionContainerEmpty) {
             return FigResponseCookies::set($response, $this->getExpirationCookie());
         }
 
-        if ($sessionContainerChanged || ($this->shouldTokenBeRefreshed($token) && ! $sessionContainer->isEmpty())) {
+        if ($sessionContainerChanged || (! $sessionContainerEmpty && $token && $this->shouldTokenBeRefreshed($token))) {
             return FigResponseCookies::set($response, $this->getTokenCookie($sessionContainer));
         }
 
@@ -263,12 +264,8 @@ final class SessionMiddleware implements MiddlewareInterface
     /**
      * {@inheritDoc}
      */
-    private function shouldTokenBeRefreshed(Token $token = null) : bool
+    private function shouldTokenBeRefreshed(Token $token) : bool
     {
-        if (null === $token) {
-            return false;
-        }
-
         if (! $token->hasClaim(self::ISSUED_AT_CLAIM)) {
             return false;
         }
@@ -317,6 +314,6 @@ final class SessionMiddleware implements MiddlewareInterface
     {
         $currentTimeProvider = $this->currentTimeProvider;
 
-        return $currentTimeProvider()->getTimestamp();
+        return $currentTimeProvider->__invoke()->getTimestamp();
     }
 }